Add EE files via upload (#198)
* Add EE files via upload

* Update EE_domestic.md

* Update EE_international.md

* Update EE_visa_quarantine.md

* Update EE_domestic.md

* Update EE_international.md

included the capitalization of "The"

* Update EE_international.md

included "Travelers who have..."

* Update EE_visa_quarantine.md

changes made according to jann's comment "This point is very difficult to understand. maybe reword it to "Self-isolation for travelers arriving in Estonia is compulsory if traveler arrives from a country that has more that 15 people per 100,000 inhabitants for the past 14 days. If traveler arrives from a country with less than 15 people per 100,000 inhabitants for the past 14 days, self-isolation is not mandatory"

* Update EE_domestic.md

* Update EE_domestic.md

* Update EE_domestic.md

* Update EE_domestic.md